rag/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
chromadb==0.5.23
sentence-transformers==3.3.1
langchain==0.3.14
langchain-community==0.3.14
langchain-text-splitters==0.3.4
pydantic==2.10.3
httpx==0.28.1
